See changelog for details
Changes to settings subsystem allowing driver/plugin code to dynamically allocate space for their settings.

Added support for STM32F446 to STM32F4xx driver.

Prepared core for laser PPI (pulses per inch) mode, plugin under development.

Bug fixes, code readbility changes, new HAL entry points and grbl core events+++

+++ drivers/SAMD21/grblHAL_MKRZERO/src/i2c.h
@@ -48,6 +48,6 @@ void I2C_GetKeycode (uint32_t i2cAddr, keycode_callback_ptr callback);
 
 #endif
 
-uint8_t *I2C_Receive (uint32_t i2cAddr, uint8_t *buf, uint32_t bytes, bool block);
-void I2C_Send (uint32_t i2cAddr, uint8_t *buf, uint8_t bytes, bool block);
-uint8_t *I2C_ReadRegister (uint32_t i2cAddr, uint8_t *buf, uint8_t bytes, bool block);
+uint8_t *I2C_Receive (uint32_t i2cAddr, uint8_t *buf, uint16_t bytes, bool block);
+void I2C_Send (uint32_t i2cAddr, uint8_t *buf, uint16_t bytes, bool block);
+uint8_t *I2C_ReadRegister (uint32_t i2cAddr, uint8_t *buf, uint16_t bytes, bool block);
